# be/docs/swagger.yaml
definitions:
  api.CitySuggestion:
    properties:
      country:
        type: string
      lat:
        type: number
      lon:
        type: number
      name:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/cities/search:
    get:
      description: Returns up to 5 city suggestions for the given query using Open-Meteo
        geocoding
      parameters:
      - description: City name (min 2 chars)
        in: query
        name: query
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.CitySuggestion'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Auto-suggest city search
      tags:
      - cities
  /api/flood/results:
    get:
      description: Returns all cached flood risk records (stubbed)
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: List all cached flood results (stub)
      tags:
      - flood
  /api/flood/risk:
    get:
      description: Returns a dynamic flood risk for given latitude and longitude
      parameters:
      - description: Latitude
        in: query
        name: latitude
        required: true
        type: string
      - description: Longitude
        in: query
        name: longitude
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get flood risk (dynamic demo)
      tags:
      - flood
  /api/weather/current:
    get:
      description: Returns the current weather for a city (live fetch, caches result)
      parameters:
      - description: City name
        in: query
        name: city
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get current weather
      tags:
      - weather
  /api/weather/result:
    get:
      description: Returns the cached/latest weather result for a city (no live fetch)
      parameters:
      - description: City name
        in: query
        name: city
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get cached weather result
      tags:
      - weather
  /api/weather/results:
    get:
      description: Returns cached weather records (city, temperature, fetched_at)
        with optional filters
      parameters:
      - description: City name
        in: query
        name: city
        type: string
      - description: Day of month
        in: query
        name: day
        type: integer
      - description: Month (1-12)
        in: query
        name: month
        type: integer
      - description: Year (e.g., 2025)
        in: query
        name: year
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: List cached weather results
      tags:
      - weather
swagger: "2.0"